Fixed error when method is defined before the class destructor
When handlinf a node that represents a method definition/decleration, the last token of the node can be the
first token of the next node.
Until now, __nodeTextTokens handled this by removing the last non-punctuation token.
When the next node represents the destructor, the last token is '~' which is a punctuation type and therefor
it was not stripped.

This parsing error does not always occur in classes because usually the constructors and destructor are
grouped together so its not common to see a method defined/declared right before the destructor.
The parsing of constructors does not depend on __nodeTextTokens stripping the last non-punctuation token.